2/16 发表文章分类

+++ src/main/java/com/gordon/blog/dao/BlogRepository.java
@@ -11,7 +11,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 public interface BlogRepository extends JpaRepository<Blog,Long>{
 
 
-    Page<Blog> findByCatalogId(int catalogId, Pageable pageable);
+    Page<Blog> findByCatalogId(Long catalogId, Pageable pageable);
 
     Page<Blog> findByUserUid(Long userUid, Pageable pageable);
